Initial commit of OE bootstrap script and basic OSD2 recipe.
This script will setup the OE environment on the user's machine.
This will allow the user to use OE to rebuild the OSD2 software
in the future.
Right now, it only allow rebuilding of a very basic image
for OE testing purpose only. It  has other packages that are not
the same as the normal build and doesn't do much.