Revert "export CC,CPP,CXX,LD variables for Makefiles"
This reverts commit 4508481b9cc09ab4d3d335b1d46b047e482d2d10.
We cannot set CC, CPP, CXX and LD because some applications require tools
to be built for x86 such as Qt4. Qt4 uses CC to build qmake, moc, and roc.
if CC=arm-linux-gcc then qmake is built with arm-linux meaning that x86
can't run it causing Qt4 build, and applications building to fail.