Use the github repositories
Complete the migration from neuros.com.cn to github.com hosting by
switching the repository URL in git-helper, so that it pull correctly
from the right place.
The ssh check is removed, and internal devs will need a special env
var to be present when using git-helper.sh to get access to private
repos.